chore(icon button): set correct color to disabled attribute
